Trait module_vault_registry_rpc::VaultRegistryRuntimeApi

pub trait VaultRegistryRuntimeApi<Block, VaultId, Balance, UnsignedFixedPoint, CurrencyId, AccountId>: Core<Block> where
    Block: Block,
    VaultId: Codec,
    Balance: Codec,
    UnsignedFixedPoint: Codec,
    CurrencyId: Codec,
    AccountId: Codec, {

    fn get_vault_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_vault_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_vaults_by_account_id(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        account_id: AccountId
    ) -> Result<Result<Vec<VaultId, Global>, DispatchError>, ApiError> { ... }

    fn get_vaults_by_account_id_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        account_id: AccountId
    ) -> Result<Result<Vec<VaultId, Global>, DispatchError>, ApiError> { ... }

    fn get_vault_total_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_vault_total_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_premium_redeem_vaults(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_premium_redeem_vaults_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_issuable_tokens(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_issuable_tokens_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_redeemable_tokens(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_vaults_with_redeemable_tokens_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError> { ... }

    fn get_issuable_tokens_from_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_issuable_tokens_from_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: VaultId, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: VaultId, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_and_collateral(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault: VaultId, 
        collateral: BalanceWrapper<Balance>, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_collateralization_from_vault_and_collateral_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault: VaultId, 
        collateral: BalanceWrapper<Balance>, 
        only_issued: bool
    ) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_wrapped(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        amount_btc: BalanceWrapper<Balance>, 
        currency_id: CurrencyId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_wrapped_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        amount_btc: BalanceWrapper<Balance>, 
        currency_id: CurrencyId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_vault(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

    fn get_required_collateral_for_vault_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vault_id: VaultId
    ) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError> { ... }

}

Provided methods

fn get_vault_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the vault’s collateral (excluding nomination)

fn get_vault_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the vault’s collateral (excluding nomination)

fn get_vaults_by_account_id(
&self,
__runtime_api_at_param__: &BlockId<Block>,
account_id: AccountId
) -> Result<Result<Vec<VaultId, Global>, DispatchError>, ApiError>

Get all the vaultIds registered by a vault’s accountId

fn get_vaults_by_account_id_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
account_id: AccountId
) -> Result<Result<Vec<VaultId, Global>, DispatchError>, ApiError>

Get all the vaultIds registered by a vault’s accountId

fn get_vault_total_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the vault’s collateral (including nomination)

fn get_vault_total_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the vault’s collateral (including nomination)

fn get_premium_redeem_vaults(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults below the premium redeem threshold, ordered in descending order of this amount

fn get_premium_redeem_vaults_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults below the premium redeem threshold, ordered in descending order of this amount

fn get_vaults_with_issuable_tokens(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults with non-zero issuable tokens, ordered in descending order of this amount

fn get_vaults_with_issuable_tokens_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults with non-zero issuable tokens, ordered in descending order of this amount

fn get_vaults_with_redeemable_tokens(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults with non-zero redeemable tokens, ordered in descending order of this amount

fn get_vaults_with_redeemable_tokens_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Result<Vec<(VaultId, BalanceWrapper<Balance>), Global>, DispatchError>, ApiError>

Get all vaults with non-zero redeemable tokens, ordered in descending order of this amount

fn get_issuable_tokens_from_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the amount of tokens a vault can issue

fn get_issuable_tokens_from_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the amount of tokens a vault can issue

fn get_collateralization_from_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: VaultId,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault

fn get_collateralization_from_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: VaultId,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault

fn get_collateralization_from_vault_and_collateral(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault: VaultId,
collateral: BalanceWrapper<Balance>,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault and collateral

fn get_collateralization_from_vault_and_collateral_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault: VaultId,
collateral: BalanceWrapper<Balance>,
only_issued: bool
) -> Result<Result<UnsignedFixedPoint, DispatchError>, ApiError>

Get the collateralization rate of a vault and collateral

fn get_required_collateral_for_wrapped(
&self,
__runtime_api_at_param__: &BlockId<Block>,
amount_btc: BalanceWrapper<Balance>,
currency_id: CurrencyId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the minimum amount of collateral required for the given amount of btc with the current threshold and exchange rate

fn get_required_collateral_for_wrapped_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
amount_btc: BalanceWrapper<Balance>,
currency_id: CurrencyId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the minimum amount of collateral required for the given amount of btc with the current threshold and exchange rate

fn get_required_collateral_for_vault(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the amount of collateral required for the given vault to be at the current SecureCollateralThreshold with the current exchange rate

fn get_required_collateral_for_vault_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vault_id: VaultId
) -> Result<Result<BalanceWrapper<Balance>, DispatchError>, ApiError>

Get the amount of collateral required for the given vault to be at the current SecureCollateralThreshold with the current exchange rate